Check for missing source on ActivityCard, refactor of contentType
